electronic_library/requirements.txt
sqlalchemy==2.0.23
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-login==0.6.3
werkzeug==3.0.1
bcrypt==4.1.2
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
reportlab==4.0.4
paramiko==3.3.1
python-dateutil==2.8.2